requirements.txt
alembic==1.0.8
aniso8601==6.0.0
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
bcrypt==3.1.6
black==18.9b0
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.0
pep8==1.7.1
pip-review==1.0
pycparser==2.19
pydocstyle==3.0.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
six==1.12.0
snowballstemmer==1.2.1
SQLAlchemy==1.3.1
toml==0.10.0
typed-ast==1.3.1
tzlocal==1.5.1
Werkzeug==0.14.1
wrapt==1.11.1
